third_party/blink/web_tests/external/wpt/app-history/per-entry-events/dispose-same-document-navigate-during.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async (t) => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(r => window.onload = () => t.step_timeout(r, 0));

  location.hash = "#1";
  location.hash = "#2";
  location.hash = "#3";

  assert_equals(appHistory.entries().length, 4);
  const [entry0, entry1, entry2, entry3] = appHistory.entries();
  assert_equals((new URL(entry2.url)).hash, "#2");
  assert_equals((new URL(entry3.url)).hash, "#3");

  let dispose3Called = 0;
  let spoonPromise;
  entry3.addEventListener("dispose", t.step_func(e => {
    ++dispose3Called;

    // This navigation will cancel the navigation to #fork that is still
    // "ongoing" (even though the URL/current entry has already been updated).
    // That is, it will cause the promise to reject (and navigateerror to fire).
    spoonPromise = appHistory.navigate("#spoon");
  }));

  await appHistory.goTo(entry1.key);

  const forkPromise = appHistory.navigate("#fork");
  assert_equals(dispose3Called, 1, "dispose for entry 3 must happen exactly once (first check)")

  // The navigation to #fork will *not* be finished by the time the navigation
  // to #spoon kicks off, so this promise will get rejected.
  await promise_rejects_dom(t, "AbortError", forkPromise);

  await spoonPromise;

  assert_equals(dispose3Called, 1, "dispose for entry 3 must happen exactly once (final check)");

  assert_equals(appHistory.entries().length, 4);
  const [finalEntry0, finalEntry1, finalEntry2, finalEntry3] = appHistory.entries();
  assert_equals(finalEntry0, entry0);
  assert_equals(finalEntry1, entry1);
  assert_not_equals(finalEntry2, entry2);
  assert_not_equals(finalEntry3, entry3);
  assert_equals(appHistory.current, finalEntry3);
  assert_equals((new URL(finalEntry2.url)).hash, "#fork");
  assert_equals((new URL(finalEntry3.url)).hash, "#spoon");
}, "navigate() during a same-document-navigation-initiated dispose works (since it's after the previous navigation)");
</script>
